[CSL-1190] Move iterators to db, remove iterKeyPrefix proxy
