+++ core/cardano-sl-core.cabal
@@ -83,7 +83,6 @@ library
                        Pos.Util.Chrono
                        Pos.Util.Config
                        Pos.Util.HVect
-                       Pos.Util.Iterator
                        Pos.Util.Limits
                        Pos.Util.LRU
                        Pos.Util.Mnemonics
