+++ src/Pos/Crypto/HDDiscovery.hs
@@ -12,9 +12,9 @@ import           Pos.Core.Types      (Address (..), addrPkDerivationPath)
 import           Pos.Crypto.HD       (HDAddressPayload, HDPassphrase, unpackHDAddressAttr)
 import           Pos.Data.Attributes (attrData)
 import           Pos.DB.Class        (MonadRealDB)
+import           Pos.DB.Iterator     (MonadIterator (..))
 import           Pos.Txp.Core        (toaOut, txOutAddress)
 import           Pos.Txp.DB          (UtxoIter, runUtxoMapIterator)
-import           Pos.Util.Iterator   (MonadIterator (..))
 
 discoverHDAddress :: MonadRealDB m => HDPassphrase -> m [(Address, [Word32])]
 discoverHDAddress walletPassphrase = safeHead <$> discoverHDAddresses [walletPassphrase]
