+++ src/Pos/Lrc/FtsPure.hs
@@ -10,12 +10,12 @@ module Pos.Lrc.FtsPure
 import qualified Data.HashMap.Strict as HM
 import           Universum
 
+import           Pos.DB.Iterator     (runListHolder)
 import           Pos.Lrc.Fts         (followTheSatoshiM)
 import           Pos.Txp.Toil.Types  (Utxo)
 import           Pos.Txp.Toil.Utxo   (utxoToStakes)
 import           Pos.Types           (Coin, SharedSeed (..), StakeholderId, coinToInteger,
                                       mkCoin, sumCoins)
-import           Pos.Util.Iterator   (runListHolder)
 
 -- | Choose several random stakeholders (specifically, their amount is
 -- currently hardcoded in 'Pos.Constants.epochSlots').
